Moved remote interaction example to separate class Added file existence checking and temporary file deletion

+++ spark_estnltk_remoteinteraction/src/spark_estnltk_remoteinteraction/HdfsRemoteInteraction.java
@@ -1,31 +1,17 @@
 package spark_estnltk_remoteinteraction;
 
 
-import java.io.BufferedInputStream;
-import java.io.BufferedOutputStream;
 import java.io.BufferedReader;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.io.OutputStream;
-import java.net.URI;
 import java.net.URISyntaxException;
-import java.security.PrivilegedExceptionAction;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.hdfs.DFSClient;
-import org.apache.hadoop.security.UserGroupInformation;
-
 import com.jcraft.jsch.ChannelExec;
 import com.jcraft.jsch.JSch;
 import com.jcraft.jsch.JSchException;
@@ -48,7 +34,7 @@ public class HdfsRemoteInteraction {
 
 
 
-	private void init() throws IOException, URISyntaxException, JSchException{
+	public void init() throws IOException, URISyntaxException, JSchException{
 //		this.conf = new Configuration();
 //		String serverUriString = "hdfs://"+host+":"+hdfsport;
 //		this.conf.set("fs.defaultFS", serverUriString);
@@ -63,7 +49,7 @@ public class HdfsRemoteInteraction {
 	}
 
 
-	/*
+	/**
 	 * Downloads the inputUri to remote local file system, converts it to Hadoop sequencefile, then copies it from local to HDFS.
 	 * returns: Path to sequencefile in HDFS if successful 
 	 */
@@ -79,38 +65,59 @@ public class HdfsRemoteInteraction {
 		System.out.println("Remote seq HDFS dest: " + inputSeqfileHdfsPath);
 		
 		String command = "";
-		int returnCode;
+		String returnValue;
 		
 		// HDFS -- check if file is in cluster hdfs
-		/*		if ( fileSystem.exists(dstPath)){
-			return "FILE_EXISTS";
-		}*/
+		command = "hadoop fs -ls " + inputSeqfileHdfsPath;
+		System.out.println("command:" + command);
+		returnValue = executeCommandOnRemote(command);
+		System.out.println("return:" + returnValue);
+		if (returnValue.length() > 0){
+			System.out.println("INFO: Seqfile already exists in HDFS, skipping to processing.");
+			return inputSeqfileHdfsPath;
+		}
+		
 		// if not:
 			//   SHELL(wget) -- download inputUri to cluster local FS
 		command = "curl " + inputUri + " --create-dirs -o " + inputTextLocalFSPath;
 		System.out.println("command:" + command);
-		returnCode = executeCommandOnRemote(command);
-		System.out.println("return:" + returnCode);
+		returnValue = executeCommandOnRemote(command);
+		System.out.println("return:" + returnValue);
+		
 			// SHELL(java) -- convert input into sequencefile 
-		//TODO: convert to seqfile command
 		String textToSeqArgString = inputTextLocalFSPath +" "+ inputTextLocalFSPath.getParent();
 		command = "java -jar "+ textToSeqJarLocation +" "+ textToSeqArgString;
 		System.out.println("command:" + command);
-		returnCode = executeCommandOnRemote(command);
-		System.out.println("return:" + returnCode);		
+		returnValue = executeCommandOnRemote(command);
+		System.out.println("return:" + returnValue);		
+		
 			//   SHELL(hdfs) --  add from cluster local to cluster HDFS
 		command = "hadoop fs -mkdir -p " + inputSeqfileHdfsPath.getParent();
 		System.out.println("command:" + command);
-		returnCode = executeCommandOnRemote(command);
-		System.out.println("return:" + returnCode);
+		returnValue = executeCommandOnRemote(command);
+		System.out.println("return:" + returnValue);
 		
 		command = "hadoop fs -put " + inputSeqfileLocalFSPath + " " + inputSeqfileHdfsPath;
 		System.out.println("command:" + command);
-		returnCode = executeCommandOnRemote(command);
-		System.out.println("return:" + returnCode);
+		returnValue = executeCommandOnRemote(command);
+		System.out.println("return:" + returnValue);
 		
+			// SHELL -- delete seqfile (and crc) from local filesystem
+		String seqFileRmString = Path.getPathWithoutSchemeAndAuthority(inputSeqfileLocalFSPath).toString();
 		
-		//TODO: delete from local filesystem
+		String crcFileRmString = Path.getPathWithoutSchemeAndAuthority(inputSeqfileLocalFSPath)
+				.getParent()+"/."+inputSeqfileLocalFSPath.getName()+".crc";
+		command = "rm " + seqFileRmString +" "+ crcFileRmString;
+		System.out.println("command:" + command);
+		returnValue = executeCommandOnRemote(command);
+		System.out.println("return:" + returnValue);
+		
+		//TODO: probably no need to run every time
+			// SHELL -- delete empty directories
+		command = "find "+ this.getLocaldir() + " -type d -empty -delete";
+		System.out.println("command:" + command);
+		returnValue = executeCommandOnRemote(command);
+		System.out.println("return:" + returnValue);
 		
 		returnMsg = "SUCCESS";
 		/*
@@ -126,62 +133,51 @@ public class HdfsRemoteInteraction {
 	}
 	
 	
-	/*
-	 * Executes command on remote server via ssh.
-	 */
-	private int executeCommandOnRemote(String command) throws JSchException, IOException{
-
-		
-		ChannelExec channelExec = (ChannelExec)session.openChannel("exec");
-		InputStream in = channelExec.getInputStream();
-		channelExec.setCommand(command);
-		channelExec.connect();
-		
-		BufferedReader reader = new BufferedReader(new InputStreamReader(in));
-		List<String> result = new ArrayList<String>();
-		String line;
-		while ((line = reader.readLine()) != null){
-			result.add(line);
-			System.out.println(line);
-		}
-		int exitStatus = channelExec.getExitStatus();
-		channelExec.disconnect();
-
-		return exitStatus;
-	}
 
 	
 	
-	/*
+	/**
 	 * Applies ESTNLTK Spark processing on the input file in HDFS.
 	 */
-	private Map<String, Path> applyProcessAndGetResultLocation(Path inputSeqfileHdfsPath, Map<String, String> estnltkOpMap, String submitParams, String taskParams) throws JSchException, IOException {
+	public Map<String, Path> applyProcessAndGetResultLocation(Path inputSeqfileHdfsPath, Map<String, String> estnltkOpMap, String submitParams, String taskParams) throws JSchException, IOException {
 		String command = "";
-		int returnCode;
+		String returnValue;
+		String suffix = ".result";
+
 		Path hdfsDirectoryPath = inputSeqfileHdfsPath.getParent();
 		String inputFileName = inputSeqfileHdfsPath.getName().substring(0, inputSeqfileHdfsPath.getName().length()-4);
-		
-		// HDFS -- check if lemmas are in cluster hdfs
+		Path sparkProcessFinalOutHdfsPath = new Path(hdfsDirectoryPath+"/"+inputFileName+suffix);
+		Map<String, Path> outputFileFinalHdfsPathMap = new HashMap<String,Path>();
+
+		// HDFS -- check if results exist in cluster hdfs
+		command = "hadoop fs -ls " + sparkProcessFinalOutHdfsPath;
+		System.out.println("command:" + command);
+		returnValue = executeCommandOnRemote(command);
+		System.out.println("return:" + returnValue);
+		if (returnValue.length() > 0){
+			System.out.println("INFO: Result file already exists.");
+			outputFileFinalHdfsPathMap.put("out",sparkProcessFinalOutHdfsPath);
+			return outputFileFinalHdfsPathMap;
+		}
 		// if not:
+		
+		
 		//   SHELL?(spark) --  run spark on input, save results to hdfs
 		Path sparkProcessOutHdfsPath = new Path(inputSeqfileHdfsPath + ".outdir");
 		String opParamString = estnltkOpMap.keySet().toString().replace(",", " ");
 		opParamString = opParamString.substring(1, opParamString.length()-1);
 		command = "spark-submit " + submitParams + " " + sparkProcessPyLocation + " "+ inputSeqfileHdfsPath +" "+ sparkProcessOutHdfsPath + " " + opParamString +" "+ taskParams;
 		System.out.println("command:" + command);
-		returnCode = executeCommandOnRemote(command);
-		System.out.println("return:" + returnCode);
+		returnValue = executeCommandOnRemote(command);
+		System.out.println("return:" + returnValue);
 		
 	
 		//   HDFS --  rename results to proper form
-		Map<String, Path> outputFileFinalHdfsPathMap = new HashMap<String,Path>();
-		String suffix = ".result";
 		Path sparkProcessTempOutHdfsPath = new Path(sparkProcessOutHdfsPath+"/part-00000");
-		Path sparkProcessFinalOutHdfsPath = new Path(hdfsDirectoryPath+"/"+inputFileName+suffix);
 		command = "hadoop fs -mv "+ sparkProcessTempOutHdfsPath +" "+ sparkProcessFinalOutHdfsPath;
 		System.out.println("command:" + command);
-		returnCode = executeCommandOnRemote(command);
-		System.out.println("return:" + returnCode);
+		returnValue = executeCommandOnRemote(command);
+		System.out.println("return:" + returnValue);
 	
 		outputFileFinalHdfsPathMap.put("out",sparkProcessFinalOutHdfsPath);
 				
@@ -190,20 +186,30 @@ public class HdfsRemoteInteraction {
 //			outputFileFinalHdfsPathMap.put(opParam, convertAndMoveResult(opDir, sparkProcessOutHdfsPath, inputFileName))
 //		);
 		
+		// HDFS -- verify result file existence
+		command = "hadoop fs -ls " + sparkProcessFinalOutHdfsPath;
+		System.out.println("command:" + command);
+		returnValue = executeCommandOnRemote(command);
+		System.out.println("return:" + returnValue);
+		if (returnValue.length() == 0){
+			System.out.println("ERROR: Result file does not exist");
+			return null;
+		}
+		
 		// HDFS --  remove out directory from hdfs
 		command = "hadoop fs -rm -r -f "+ sparkProcessOutHdfsPath;
 		System.out.println("command:" + command);
-		returnCode = executeCommandOnRemote(command);
-		System.out.println("return:" + returnCode);
+		returnValue = executeCommandOnRemote(command);
+		System.out.println("return:" + returnValue);
 		
 		// return location(s) map of results
 		return outputFileFinalHdfsPathMap;
 	}
 
-	/*
+	/**
 	 * Moves and renames spark output files to the correct location. 
 	 */
-	private Path convertAndMoveResult (String opDir, Path hdfsDirectoryPath, String inputFileName){
+	public Path convertAndMoveResult (String opDir, Path hdfsDirectoryPath, String inputFileName){
 		String outputFileName = "part-00000";
 		Path outputFileTempHdfsPath = new Path(hdfsDirectoryPath +"/"+ opDir +"/"+ outputFileName);
 		String outputFinalSuffix = opDir;
@@ -211,8 +217,8 @@ public class HdfsRemoteInteraction {
 		try{
 			String command = "hadoop fs -mv "+ outputFileTempHdfsPath +" "+ outputFileFinalHdfsPath;
 			System.out.println("command:" + command);
-			int returnCode = executeCommandOnRemote(command);
-			System.out.println("return:" + returnCode);
+			String returnValue = executeCommandOnRemote(command);
+			System.out.println("return:" + returnValue);
 			return outputFileFinalHdfsPath;
 		}
 		catch (Exception e){
@@ -221,148 +227,165 @@ public class HdfsRemoteInteraction {
 		}
 	}
 	
-	//XXX: TESTING ONLY
-	public void runFullProcess() throws JSchException, IOException, InterruptedException{
-		// TODO: get from config
-		this.localdir = "";
-		this.hdfsdir = "";
-		this.host = "";
-		this.hdfsport = 8020;
-		this.sshport = 22;
-		this.username = "";
-		this.password = "";
-		this.textToSeqJarLocation = "";
-		this.sparkProcessPyLocation = "";
+	/**
+	 * Executes command on remote server via ssh.
+	 */
+	public String executeCommandOnRemote(String command) throws JSchException, IOException{
+		ChannelExec channelExec = (ChannelExec)session.openChannel("exec");
+		InputStream in = channelExec.getInputStream();
+		channelExec.setCommand(command);
 		
-			//watch out for script injection (purify URI beforehand)
-			//remove trailing "/"-symbol
-//		String inputUri = "http://textfiles.com/games/abbrev.txt";
-		String inputUri = "http://www.cs.ut.ee/et/teadus/uurimisruhmad";
-		String cleanedInputUri = inputUri.replace("//", "/").replace(":", "_").replace("?", "_").replace("=", "_");
-		String localFSDest = "file://" + localdir + "/" + cleanedInputUri;
-		String hdfsDest = "hdfs://"+ hdfsdir + "/" + cleanedInputUri;
-		String type = "html";
-		String submitParams = "--master yarn-cluster --num-executors 4";
-		String taskParams = "";
-		Map<String,String> estnltkOpMap = new HashMap<String,String>();
-		estnltkOpMap.put("-lemma","lemma");
-		estnltkOpMap.put("-ner","ner");
-		
-		// construct Spark process params
-		if (type != "html"){
-			taskParams += " -isPlaintextInput";
+		BufferedReader reader = new BufferedReader(new InputStreamReader(in));
+
+		channelExec.connect();
+				
+		String result = "";
+		String line;
+		while ((line = reader.readLine()) != null){
+			result+=line;
+//			System.out.println("RET:"+line);
 		}
+		reader.close();
+		int exitStatus = channelExec.getExitStatus();
+		channelExec.disconnect();
 		
-		// init class fields
-		try {
-			init();
-		} catch (IOException e1) {
-			e1.printStackTrace();
-		} catch (URISyntaxException e) {
-			e.printStackTrace();
-		}
-
-		// Save from input URI to HDFS, get HDFS location
-		Path hdfsPath = null;		
-		session.connect();
-
-		hdfsPath = saveFromUriToRemoteHdfs(inputUri, localFSDest, hdfsDest);
-
-		System.out.println("SEQFILE LOC: " + hdfsPath);
-
-		// Apply ESTNLTK processes to seqfile, get result location
-		Map<String, Path> processResultLocationsMap = applyProcessAndGetResultLocation(hdfsPath, estnltkOpMap, submitParams, taskParams);
-		System.out.println("OUTMAP: "+ processResultLocationsMap);
-		session.disconnect(); //closes session, needs reinitializing before reopen?
+		return result;
 		
+//		System.out.println("RES:"+result.length());
+//		if (result.length() > 0){
+//			return result;
+//		}
+//		else{
+//			return String.valueOf(exitStatus);
+//		}
 	}
 
-
-
+	/**
+	 * Server local file system directory used as a buffer to download the target URI content into, then converted into Hadoop Sequencefile, then uploaded to HDFS. 
+	 */
 	public String getLocaldir() {
 		return localdir;
 	}
 
-
+	/**
+	 * Server local file system directory used as a buffer to download the target URI content into, then converted into Hadoop Sequencefile, then uploaded to HDFS. 
+	 */
 	public void setLocaldir(String localdir) {
 		this.localdir = localdir;
 	}
 
-
+	/**
+	 * IP of server.
+	 */
 	public String getHost() {
 		return host;
 	}
 
-
+	/**
+	 * IP of server.
+	 */
 	public void setHost(String host) {
 		this.host = host;
 	}
 
-
+	/**
+	 * Port of SSH connection to server.
+	 */
 	public int getSshport() {
 		return sshport;
 	}
 
-
+	/**
+	 * Port of SSH connection to server.
+	 */
 	public void setSshport(int sshport) {
 		this.sshport = sshport;
 	}
 
-
+	/**
+	 * Server user name.
+	 */
 	public String getUsername() {
 		return username;
 	}
 
-
+	/**
+	 * Server user name.
+	 */
 	public void setUsername(String username) {
 		this.username = username;
 	}
 
-
+	/**
+	 * Password of server user.
+	 */
 	public String getPassword() {
 		return password;
 	}
 
-
+	/**
+	 * Password of server user.
+	 */
 	public void setPassword(String password) {
 		this.password = password;
 	}
 
-
+	/**
+	 * Directory on Hadoop File System (HDFS) where sequencefiles and processing results are stored. Specified directory is used as a root directory, subdirectories are created for each different input.
+	 */
 	public String getHdfsdir() {
 		return hdfsdir;
 	}
 
-
+	/**
+	 * Directory on Hadoop File System (HDFS) where sequencefiles and processing results are stored. Specified directory is used as a root directory, subdirectories are created for each different input.
+	 */
 	public void setHdfsdir(String hdfsdir) {
 		this.hdfsdir = hdfsdir;
 	}
 
-
+	/**
+	 * Path on the server local file system to TextToSeqfile.java, used to convert textfiles into sequence files for processing.
+	 */
 	public String getTextToSeqJarLocation() {
 		return textToSeqJarLocation;
 	}
 
-
+	/**
+	 * Path on the server local file system to TextToSeqfile.java, used to convert textfiles into sequence files for processing.
+	 */
 	public void setTextToSeqJarLocation(String textToSeqJarLocation) {
 		this.textToSeqJarLocation = textToSeqJarLocation;
 	}
 
-
+	/**
+	 * Path on the server local file system to process.py, which executes ESTNLTK processes in Spark. 
+	 */
 	public String getSparkProcessPyLocation() {
 		return sparkProcessPyLocation;
 	}
 
-
+	/**
+	 * Path on the server local file system to process.py, which executes ESTNLTK processes in Spark. 
+	 */
 	public void setSparkProcessPyLocation(String sparkProcessPyLocation) {
 		this.sparkProcessPyLocation = sparkProcessPyLocation;
 	}
 
-
-	public static void main(String[] args) throws JSchException, IOException, InterruptedException{
-		HdfsRemoteInteraction interactor = new HdfsRemoteInteraction();
-		interactor.runFullProcess();
+	/**
+	 * JSch shell connection session to the server.
+	 */
+	public Session getSession() {
+		return session;
 	}
+
+	/**
+	 * JSch shell connection session to the server.
+	 */
+	public void setSession(Session session) {
+		this.session = session;
+	}
+
 
 }
